chore: refactor I18n types outside of window object
to keep ts/localization folder self contained
